+++ etc/acmg.db.list.txt
@@ -1,12 +1,12 @@
-PS1PM5.MutationName.count	PS1PM5.MutationName.count
-PS1PM5.pHGVS1.count	PS1PM5.pHGVS1.count
-PS1PM5.AApos.count	PS1PM5.AApos.count
-PM1InterproDomain	PM1.InterproDomain.count
-PM1PfamIdDomain	PM1.pfamId.count
-LOFList	PVS1GeneList.json
-transcriptInfo	transcript.info.json
-PathogenicLite	PathogenicLite.bed.gz
-PP2GeneList	PP2.gene.list
-LateOnset	BS2.gene.list
-BP1GeneList	BP1.gene.list
-BA1ExceptionList	BA1.exception.list
+PS1PM5.MutationName.count	acmg/PS1PM5.MutationName.count
+PS1PM5.pHGVS1.count	acmg/PS1PM5.pHGVS1.count
+PS1PM5.AApos.count	acmg/PS1PM5.AApos.count
+PM1InterproDomain	acmg/PM1.InterproDomain.count
+PM1PfamIdDomain	acmg/PM1.pfamId.count
+LOFList	acmg/PVS1GeneList.json
+transcriptInfo	acmg/transcript.info.json
+PathogenicLite	acmg/PathogenicLite.bed.gz
+PP2GeneList	acmg/PP2.gene.list
+LateOnset	acmg/BS2.gene.list
+BP1GeneList	acmg/BP1.gene.list
+BA1ExceptionList	acmg/BA1.exception.list
